ai-service/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Computer Vision
opencv-python-headless==4.10.0.84
numpy>=1.26.0

# AI / ML
ultralytics==8.4.14
requests>=2.31.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.18
Pillow>=10.0.0
yt-dlp>=2026.2.0
